+++ src/id_translation/offline/types.py
@@ -31,6 +31,8 @@ class PlaceholderTranslations(_t.Generic[_tt.SourceType]):
     """Matrix of shape `N x M` where `N` is the number of IDs returned and `M` is the length of `placeholders`."""
     id_pos: int = -1
     """Position if the ID placeholder in `placeholders`."""
+    placeholder_aliases: dict[str, str] = _dataclasses.field(default_factory=dict)
+    """Alternative :attr:`placeholder <placeholders>` names."""
 
     @classmethod
     def make(cls, source: _tt.SourceType, data: "MakeTypes[_tt.SourceType, _tt.IdTypes]") -> _t.Self: